클래스 파일로 분리하고 디렉터리 생성
