+++ KoreaTripRank/KoreaTripRank/Model/DataModel.swift
@@ -169,3 +169,99 @@ class AreaDatabase {
 }
 
 
+// MARK: - 위도 경도 변환
+// 위도 경도 변환 데이터 모델
+struct ConvertedLocationModel {
+    public var lat: Double
+    public var lng: Double
+    public var x: Int?
+    public var y: Int?
+    
+    init(lat: Double, lng: Double, x: Int? = nil, y: Int? = nil) {
+        self.lat = lat
+        self.lng = lng
+        self.x = x
+        self.y = y
+    }
+    
+    mutating func convertGRID_GPS(mode: Int, lat_X: Double, lng_Y: Double) {
+        let RE = 6371.00877 // 지구 반경(km)
+        let GRID = 5.0 // 격자 간격(km)
+        let SLAT1 = 30.0 // 투영 위도1(degree)
+        let SLAT2 = 60.0 // 투영 위도2(degree)
+        let OLON = 126.0 // 기준점 경도(degree)
+        let OLAT = 38.0 // 기준점 위도(degree)
+        let XO:Double = 43 // 기준점 X좌표(GRID)
+        let YO:Double = 136 // 기1준점 Y좌표(GRID)
+        let DEGRAD = Double.pi / 180.0
+        let RADDEG = 180.0 / Double.pi
+        let re = RE / GRID
+        let slat1 = SLAT1 * DEGRAD
+        let slat2 = SLAT2 * DEGRAD
+        let olon = OLON * DEGRAD
+        let olat = OLAT * DEGRAD
+        let TO_GRID = 0
+        let TO_GPS = 1
+        
+        var sn = tan(Double.pi * 0.25 + slat2 * 0.5) / tan(Double.pi * 0.25 + slat1 * 0.5)
+        sn = log(cos(slat1) / cos(slat2)) / log(sn)
+        var sf = tan(Double.pi * 0.25 + slat1 * 0.5)
+        sf = pow(sf, sn) * cos(slat1) / sn
+        var ro = tan(Double.pi * 0.25 + olat * 0.5)
+        ro = re * sf / pow(ro, sn)
+        var rs = ConvertedLocationModel(lat: 0, lng: 0, x: 0, y: 0)
+        
+        if mode == TO_GRID {
+            rs.lat = lat_X
+            rs.lng = lng_Y
+            var ra = tan(Double.pi * 0.25 + (lat_X) * DEGRAD * 0.5)
+            ra = re * sf / pow(ra, sn)
+            var theta = lng_Y * DEGRAD - olon
+            if theta > Double.pi {
+                theta -= 2.0 * Double.pi
+            }
+            if theta < -Double.pi {
+                theta += 2.0 * Double.pi
+            }
+            
+            theta *= sn
+            rs.x = Int(floor(ra * sin(theta) + XO + 0.5))
+            rs.y = Int(floor(ro - ra * cos(theta) + YO + 0.5))
+        }
+        else {
+            rs.x = Int(lat_X)
+            rs.y = Int(lng_Y)
+            let xn = lat_X - XO
+            let yn = ro - lng_Y + YO
+            var ra = sqrt(xn * xn + yn * yn)
+            if (sn < 0.0) {
+                ra = -ra
+            }
+            var alat = pow((re * sf / ra), (1.0 / sn))
+            alat = 2.0 * atan(alat) - Double.pi * 0.5
+            
+            var theta = 0.0
+            if (abs(xn) <= 0.0) {
+                theta = 0.0
+            }
+            else {
+                if (abs(yn) <= 0.0) {
+                    theta = Double.pi * 0.5
+                    if (xn < 0.0) {
+                        theta = -theta
+                    }
+                }
+                else {
+                    theta = atan2(xn, yn)
+                }
+            }
+            let alon = theta / sn + olon
+            rs.lat = alat * RADDEG
+            rs.lng = alon * RADDEG
+        }
+        self.lat = rs.lat
+        self.lng = rs.lng
+        self.x = rs.x
+        self.y = rs.y
+    }
+}
